move APMTracer to tracing so that only APM is exported
