+++ modules/apm/src/test/java/org/elasticsearch/telemetry/apm/internal/APMAgentSettingsTests.java
@@ -6,9 +6,10 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.telemetry.apm;
+package org.elasticsearch.telemetry.apm.internal;
 
 import org.elasticsearch.common.settings.Settings;
+import org.elasticsearch.telemetry.apm.settings.APMAgentSettings;
 import org.elasticsearch.test.ESTestCase;
 
 import static org.mockito.Mockito.spy;
